Group everything templating-related in utils file.
